adk-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
google-adk>=1.0.0
google-cloud-aiplatform[adk,agent_engines]>=1.90.0
python-multipart
sse-starlette